Fix additional hours int to float

+++ react_app/src/views/AssignmentView.js
@@ -365,7 +365,7 @@ export class ModalAssignmentMassActions extends Component{
         for (let item of this.state.data.assignments){
             let add = {};
             add.id = item.id;
-            add.nbAdditionalHours = parseInt(this.state.nbAdditionalHours);
+            add.nbAdditionalHours = parseFloat(this.state.nbAdditionalHours);
             add.additionalHoursReason = this.state.additionalHoursReason;
             add.templateId = this.state.data.template.id;
             newItems.push(add);
